Refactor: move utility functions for determining whether use case is global and getting use case statuses to utils file
