+++ src/components/MAPI/permissions/PermissionsOverview/PermissionsOverview.tsx
@@ -2,12 +2,15 @@ import { selectOrganizations } from 'model/stores/organization/selectors';
 import React, { useMemo, useState } from 'react';
 import { useSelector } from 'react-redux';
 import { Tab, Tabs } from 'UI/Display/Tabs';
-import { cartesian } from 'utils/helpers';
 
 import PermissionsUseCases from '../PermissionsUseCases';
-import { IPermissionsUseCase } from '../types';
+import { IPermissionsUseCase, PermissionsUseCaseStatuses } from '../types';
 import { usePermissions } from '../usePermissions';
-import { getPermissionsUseCases, hasPermission } from '../utils';
+import {
+  getPermissionsUseCases,
+  getStatusesForUseCases,
+  isGlobalUseCase,
+} from '../utils';
 
 interface IGlobalPermissionsOverviewProps {
   useCases: IPermissionsUseCase[];
@@ -42,41 +45,10 @@ const OrganizationsPermissionsOverview: React.FC<
     (a?.name || a.id).localeCompare(b?.name || b.id)
   );
 
-  const useCasesStatuses: Record<
-    string,
-    Record<string, boolean>
-  > = useMemo(() => {
-    if (typeof permissions === 'undefined') {
-      return {};
-    }
+  const useCasesStatuses: PermissionsUseCaseStatuses = useMemo(() => {
+    if (typeof permissions === 'undefined') return {};
 
-    const statuses: Record<string, Record<string, boolean>> = {};
-    useCases.forEach((useCase) => {
-      const useCasePermissions = useCase.permissions.flatMap((permission) =>
-        cartesian(permission.verbs, permission.resources, permission.apiGroups)
-      );
-
-      statuses[useCase.name] = {};
-      sortedOrganizations.forEach((org) => {
-        const permissionsValues = useCasePermissions.map((p) => {
-          const [verb, resource, apiGroup] = p as string[];
-
-          return hasPermission(
-            permissions,
-            org.namespace ?? '',
-            verb,
-            apiGroup,
-            resource
-          );
-        });
-
-        statuses[useCase.name][org.id] = permissionsValues.every(
-          (v) => v === true
-        );
-      });
-    });
-
-    return statuses;
+    return getStatusesForUseCases(permissions, useCases, sortedOrganizations);
   }, [permissions, useCases, sortedOrganizations]);
 
   if (typeof permissions === 'undefined') {
@@ -92,10 +64,6 @@ const OrganizationsPermissionsOverview: React.FC<
   );
 };
 
-const isGlobal = (useCase: IPermissionsUseCase) =>
-  (useCase.scope.namespaces && useCase.scope.namespaces[0] === 'default') ||
-  useCase.scope.cluster;
-
 const PermissionsOverview: React.FC = () => {
   const [activeTab, setActiveTab] = useState(0);
 
@@ -105,9 +73,9 @@ const PermissionsOverview: React.FC = () => {
     return null;
   }
 
-  const globalUseCases = useCases.filter((useCase) => isGlobal(useCase));
+  const globalUseCases = useCases.filter((useCase) => isGlobalUseCase(useCase));
   const organizationsUseCases = useCases.filter(
-    (useCase) => !isGlobal(useCase)
+    (useCase) => !isGlobalUseCase(useCase)
   );
 
   return (
